load() and load_multiles() support BinaryIO objects. Helper functions to equate array and stack are moved from tools to image module for typing support
