+++ src/Dosepy/image.py
@@ -6,7 +6,8 @@ DESCRIPTION
     This module holds functionalities for tif image loading and manipulation.
     ArryaImage class is used as representation of dose distributions.
     The content is heavily based from 
-    `pylinac <https://pylinac.readthedocs.io/en/latest/_modules/pylinac/core/image.html>`_, and `omg_dosimetry <https://omg-dosimetry.readthedocs.io/en/latest/>`_
+    `pylinac <https://pylinac.readthedocs.io/en/latest/_modules/pylinac/core/image.html>`_,
+    and `omg_dosimetry <https://omg-dosimetry.readthedocs.io/en/latest/>`_
 
 """
 
@@ -16,6 +17,12 @@ import os.path as osp
 from typing import Any, Union, BinaryIO
 import imageio.v3 as iio
 from abc import ABC, abstractmethod
+
+import copy
+import math
+import logging
+
+from pathlib import Path
 
 import matplotlib.pyplot as plt
 import matplotlib.patches as mpatches
@@ -28,17 +35,27 @@ from skimage.measure import label, regionprops
 from skimage.filters.rank import mean
 from skimage.transform import rotate
 from .tools.resol import equate_resolution
-from .tools.files_to_image import equate_array_size, average_tiff_images, stack_images
-from .tools.array_utils import filter_array
 
-import math
+from .tools.array_utils import filter_array
 
 from .i_o import retrieve_dicom_file, is_dicom_image
 
 MM_PER_INCH = 25.4
 MIN_AREA_FOR_FILMS = 400  # 20x20 mm^2
 
+FILE_TYPE = "file"
+STREAM_TYPE = "stream"
+
 ImageLike = Union["ArrayImage", "TiffImage"]
+
+
+logging.basicConfig(
+    format="{asctime} - {levelname} - {message}",
+    style="{",
+    datefmt="%Y-%m-%d %H:%M",
+    level=logging.WARNING,
+    filename="image_module.log",
+    )
 
 
 def load(path: str | Path | np.ndarray | BinaryIO, **kwargs) -> ImageLike:
@@ -103,7 +120,7 @@ def load(path: str | Path | np.ndarray | BinaryIO, **kwargs) -> ImageLike:
                         a valid file.")
 
 
-def load_multiples(files: list[str]) -> ImageLike:
+def load_multiples(files: list[str | Path | BinaryIO]) -> ImageLike:
     """
     Load multiple images into a single one. 
     Equate TIFF files to have the same array size.
@@ -131,7 +148,7 @@ def load_multiples(files: list[str]) -> ImageLike:
         
         img = load(files[0]) # Placeholder
         equated_images = equate_array_size(images, axis=("width", "height"))
-        averaged_images = average_tiff_images(files, equated_images)
+        averaged_images = average_tiff_images(equated_images)
         stacked = stack_images(averaged_images, padding=6)
         img.array = stacked.array
         return img
@@ -170,6 +187,8 @@ class BaseImage(ABC):
 
     array: np.ndarray
     path: str | Path
+    base_path: str  # Name of the file
+    source: str
 
     def __init__(self, path: str | Path | np.ndarray):
         """
@@ -179,11 +198,18 @@ class BaseImage(ABC):
             The path to the image.
         """
 
+        # Check for a file
         if isinstance(path, (str, Path)) and not osp.isfile(path):
             raise FileExistsError(
                 f"File `{path}` does not exist. Verify the file path name.")
-        else:
+        
+        elif isinstance(path, (str, Path)) and osp.isfile(path):
             self.path = path
+            self.base_path = osp.basename(path)
+            self.source = FILE_TYPE
+        
+        else:
+            self.source = STREAM_TYPE
 
         super().__init__()
 
@@ -315,7 +341,7 @@ class TiffImage(BaseImage):
 
     def __init__(
         self,
-        path: str | Path,
+        path: str | Path | BinaryIO,
         *,
         dpi: float | None = None,
         ):
@@ -333,13 +359,18 @@ class TiffImage(BaseImage):
             Label image regions.
         """
         super().__init__(path)
-        self.props = iio.improps(path)
-        self.array = iio.imread(path)
+        self.props = iio.improps(path, extension=".tif")
+        if not self.props:
+            logging.WARNING("Was not possible to read image props")
+            print("Was not possible to read image props")
+        self.array = iio.imread(path, extension=".tif")
 
         try:
             dpi = self.props.spacing[0]
 
         except:
+            logging.WARNING("Image props has no spacing props.")
+            print("Image props has no spacing props.")
             pass
 
         self._dpi = dpi
@@ -1121,3 +1152,243 @@ class DoseImage(ArrayImage):
         self._reference_point  = reference_point
         self._orientation = orientation
         self._dose_unit = dose_unit
+
+
+def equate_array_size(
+        image_list: list,
+        axis: tuple[str, ...] = ("height", "width"),
+        ) -> list:
+    """
+    Equate TIFF files to have the same array size with respect of the smallest one.
+    Pixels are cropped equally from both sides.
+
+    Parameters
+    ----------
+    image_list : list
+        List with images (TiffImage, ArrayImage instance).
+
+    axis : str
+        Axis to equate: height, width or both.
+
+    Return
+    ------
+        A list with the new images.
+    """
+    
+    cropped_images = copy.deepcopy(image_list)
+    idx_min_height, idx_min_width = _find_smallest_image(image_list)
+
+    if "height" in axis:
+        for count, img in enumerate(image_list):
+            if count == idx_min_height: continue
+            cropped_images[count] = _equate_height(image_list[idx_min_height], img)
+            
+    image_list = cropped_images
+    if "width" in axis:
+
+        for count, img in enumerate(image_list):
+            if count == idx_min_width: continue
+            cropped_images[count] = _equate_width(image_list[idx_min_width], img)
+
+    return cropped_images
+
+
+def average_tiff_images(images: list[TiffImage | ArrayImage]) -> list:
+    """
+    Average images with the same file name, ignoring last 7 characters.
+
+    Parameters
+    ----------
+    paths : list
+        list of strings with the tiff file path.
+
+    images : list
+        list of TiffImage or ArrayImage objects.
+
+    Return
+    ------
+    averaged_images : list
+        list of TiffImage or ArrayImage objects.
+
+    Note
+    ----
+    Since the last 7 characters of the file name are ignored, the function
+    average the next files:
+
+    - my_file_name_001.tif
+    - my_file_name_002.tif
+    - my_file_name_003.tif
+
+    """
+
+    # Get base_name for identification
+    paths_as_str = [img.base_path for img in images]
+
+    # Create a list with no duplicate names
+    unique_names = list(set([file[:-7] for file in paths_as_str]))
+    
+    averaged_images = []
+    for unique in unique_names:
+        to_merge =[]
+        buff = copy.deepcopy(images[0])
+
+        # Catch files with same name
+        for file_name, image in zip(paths_as_str, images):
+            if file_name[:-7] == unique:
+                to_merge.append(image)
+        
+        new_array = np.stack(tuple(img.array for img in to_merge), axis=-1)
+        buff.array = np.mean(new_array, axis=3)
+        averaged_images.append(buff)
+    
+    return averaged_images
+
+
+def stack_images(img_list: list, axis=0, padding=0):
+    """
+    Takes in a list of images and concatenate them side by side.
+    Useful for film calibration, when more than one image is needed
+    to scan all gafchromic bands.
+    
+    Adapted from OMG_Dosimetry (https://omg-dosimetry.readthedocs.io/en/latest/)
+
+    Parameters
+    ----------
+    img_list : list
+        The images to be stacked. List of TiffImage or ArrayImage.
+
+    axis : int, default: 0
+        The axis along which the arrays will be joined. 0 if vertical or 1 if horizontal.
+
+    padding : float, default: 0
+        Add padding in milimeters to simulate an empty space betwen films.
+
+    Returns
+    -------
+    ::class:`~Dosepy.image.TiffImage`
+        Instance of a TiffImage class.
+
+    Example
+    -------
+
+        >>> img1 = load(np.ones((5, 5, 3)), dpi=1)
+        >>> img2 = load(np.ones((5, 5, 3)), dpi=1)
+
+        >>> img = stack_images([img1, img2])
+
+        >>> img.shape  # (10, 5, 3)
+    """
+
+    first_img = copy.deepcopy(img_list[0])
+
+    # Check that all images are the same width
+    for img in img_list:
+        
+        if axis == 0:
+            if img.shape[1] != first_img.shape[1]:
+                raise ValueError("Images were not the same width")
+        if axis == 1:
+            if img.shape[0] != first_img.shape[0]:
+                raise ValueError("Images were not the same height")
+
+    #height = first_img.shape[0]
+    width = first_img.shape[1]
+
+    padding_pixels = int(padding * img_list[0].dpmm)
+
+    new_img_list = []
+    
+    for img in img_list:
+
+        height = img.shape[0]
+
+        background = np.zeros(
+            (2*padding_pixels + height, 2*padding_pixels + width, 3)
+            ) + int(2**16 - 1)
+
+        background[
+            padding_pixels: padding_pixels + height,
+            padding_pixels: padding_pixels + width,
+            :
+            ] = img.array
+        new_img = copy.deepcopy(img)
+        new_img.array = background
+        new_img_list.append(new_img)
+    
+    new_array = np.concatenate(tuple(img.array for img in new_img_list), axis)
+    first_img.array = new_array
+
+    return first_img
+
+def _find_smallest_image(images: list[ImageLike]):
+
+    min_higth = images[0].shape[0]
+    min_width = images[0].shape[1]
+
+    index_min_height = 0
+    index_min_width = 0
+
+    for count, img in enumerate(images[1:], start=1):
+
+        if img.shape[0] < min_higth:
+            min_higth = img.shape[0]
+            index_min_height = count
+
+        if img.shape[1] < min_width:
+            min_width = img.shape[1]
+            index_min_width = count
+
+    logging.debug(f"Smallest height: {min_higth} at index {index_min_height}")
+    logging.debug(f"Smallest width: {min_width} at index {index_min_width}")
+    return index_min_height, index_min_width
+
+
+def _equate_height(small_image, image):
+    """
+    Crop the image to have the same height as the small_image. 
+    If the difference is odd, the extra pixel is cropped from the top.
+    Otherwise, the extra pixels are cropped equally from both sides.
+    """
+    logging.debug(f"Image height before cropping: {image.shape}")
+    height_diff = abs(int(image.shape[0] - small_image.shape[0]))
+    logging.debug(f"Height difference: {height_diff}")
+
+    if height_diff > 0:
+                
+        if height_diff == 1:
+            image.crop(height_diff, edges="bottom")
+
+        elif not(height_diff%2):
+            image.crop(int(height_diff/2), edges=('bottom', 'top'))
+
+        else:
+            image.crop(int(math.floor(height_diff/2)), edges="top")
+            image.crop(int(math.floor(height_diff/2) + 1), edges="bottom")
+
+
+    logging.debug(f"Image height after cropping: {image.shape}")
+    return image
+
+
+def _equate_width(small_image, image):
+    """
+    Crop the image to have the same width as the small_image.
+    If the difference is odd, the extra pixel is cropped from the left.
+    Otherwise, the extra pixels are cropped equally from both sides.
+    """
+
+    width_diff = abs(int(image.shape[1] - small_image.shape[1]))
+
+    if width_diff > 0:
+
+        if width_diff==1:
+            image.crop(width_diff, edges="right")
+
+        elif not(width_diff%2):
+            image.crop(int(width_diff/2), edges=("left", "right"))
+
+        else:
+            image.crop(int(math.floor(width_diff/2)), edges="left")
+            image.crop(int(math.floor(width_diff/2) + 1), edges="right")
+
+    return image
